Enabled the deck seed for each round to be stored
To enable replaying all of the actions for a round we need to be able to reconstitute
the set of shuffled cards - to do that we store the seed used to shuffle the cards.
Unfortunately Random.new_seed generates a magical Integer that cannot be easily
and reliably stored in an int(8) column so added a module to generate a bigint safe
random number to use as a seed.