Enabled the game to detect and display when a round is over
Round over detection is based on the stock being exhausted.
Added detection code to game engine and then propogate the flag through presenters to the view.
View will render a message and start round form if the game is over and no player actions form.

NOTE: None of the actions on game engine currently check to see if the round is over